Always parse target duration as integer
Following the HLS spec, target duration is an integer
value.

+++ m3u8/parser.py
@@ -452,7 +452,7 @@ def _parse_byterange(line, state, **kwargs):
 
 
 def _parse_targetduration(**parse_kwargs):
-    return _parse_simple_parameter(cast_to=float, **parse_kwargs)
+    return _parse_simple_parameter(cast_to=int, **parse_kwargs)
 
 
 def _parse_media_sequence(**parse_kwargs):
